Extract ACPI and SMBIOS board emissions
