+++ crates/fstart-codegen/src/stage_gen/board_gen.rs
@@ -58,6 +58,7 @@ use fstart_types::{
 use super::tokens::hex_addr;
 
 mod boot_media;
+mod caps_tables;
 mod lifecycle;
 mod logger;
 mod model;
@@ -65,6 +66,7 @@ mod phases;
 mod state;
 
 use boot_media::{anchor_bytes_stmt, match_boot_media};
+use caps_tables::{acpi_load_body, acpi_prepare_body, memory_detect_body, smbios_prepare_body};
 use lifecycle::{init_all_devices_body, init_device_body};
 use logger::install_logger_body;
 use model::{device_provides, BoardCtx};
@@ -1267,311 +1269,6 @@ fn late_driver_init_body(ctx: &BoardCtx<'_>) -> TokenStream {
         #(#southbridge_finalize)*
         #(#mainboard_finalize)*
     }
-}
-
-/// Emit the body of `Board::acpi_load`.
-///
-/// One arm per enabled device providing the `AcpiTableProvider`
-/// service.  Each arm:
-///
-/// 1. Defines a method-local `static _ACPI_LOAD_BUF: UnsafeCell<[u8;
-///    256 * 1024]>` (256 KiB is enough for QEMU Q35's ~128 KiB of
-///    tables plus RSDP + alignment).  The buffer is leaked after
-///    the call so tables persist for the OS.
-/// 2. Calls `fstart_capabilities::acpi_load(self.<field>.as_ref()?,
-///    buf, <name>)` and writes the returned RSDP address to
-///    `self._acpi_rsdp_addr` — read later by `payload_load` / a
-///    future `acpi_prepare`.
-/// 3. On success returns `Ok(())`; on error logs + halts.
-///
-/// Wildcard arm: log + halt.  Boards with no `AcpiTableProvider`
-/// device reduce to just the wildcard, which is dead code (no
-/// `AcpiLoad` in their capability list).
-///
-/// # Note on the buffer placement
-///
-/// A 256 KiB buffer is too large to live on the stack (firmware
-/// stacks are typically 16–64 KiB) or in the `_BoardDevices`
-/// struct (we want the struct small enough to live on the stack).
-/// Each method gets its own `static` via `UnsafeCell<[u8; N]>` —
-/// matches the old codegen's approach.  Because the buffer is
-/// scoped to the method, it does not collide with other statics
-/// and needs no per-device suffix.
-fn acpi_load_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    let arms = enabled_indices(ctx.devices, ctx.instances, ctx.excluded)
-        .filter(|idx| device_provides(ctx, *idx, Service::AcpiTableProvider))
-        .map(|idx| {
-            let dev = &ctx.devices[idx];
-            let field = format_ident!("{}", dev.name.as_str());
-            let id_lit = proc_macro2::Literal::u8_unsuffixed(idx as u8);
-            let dev_name = dev.name.as_str();
-            quote! {
-                #id_lit => {
-                    // 256 KiB buffer for ACPI tables — persists for
-                    // the OS lifetime.  Wrapped in a newtype with
-                    // `unsafe impl Sync` because `UnsafeCell` alone
-                    // is not `Sync` and cannot live in a `static`.
-                    #[repr(align(16))]
-                    struct _AcpiLoadBufStore(core::cell::UnsafeCell<[u8; 256 * 1024]>);
-                    // SAFETY: single-threaded firmware init, buffer
-                    // used exactly once (the executor only dispatches
-                    // `AcpiLoad` for one id per stage).
-                    unsafe impl Sync for _AcpiLoadBufStore {}
-                    static _ACPI_LOAD_BUF: _AcpiLoadBufStore =
-                        _AcpiLoadBufStore(core::cell::UnsafeCell::new([0u8; 256 * 1024]));
-                    let _acpi_buf = unsafe { &mut *_ACPI_LOAD_BUF.0.get() };
-                    let rsdp = fstart_capabilities::acpi_load(
-                        self.#field
-                            .as_ref()
-                            .unwrap_or_else(|| fstart_platform::halt()),
-                        _acpi_buf,
-                        #dev_name,
-                    )
-                    .map_err(|_| fstart_services::device::DeviceError::InitFailed)?;
-                    self._acpi_rsdp_addr = rsdp;
-                    Ok(())
-                }
-            }
-        });
-    quote! {
-        match id {
-            #(#arms)*
-            _ => {
-                // Executor contract violation — `acpi_load(id)` is
-                // only dispatched for ids the plan marked as
-                // `AcpiTableProvider` providers.
-                fstart_log::error!("acpi_load: unknown device id {}", id);
-                fstart_platform::halt();
-            }
-        }
-    }
-}
-
-/// Emit the body of `Board::memory_detect`.
-///
-/// One arm per enabled device providing the `MemoryDetector`
-/// service.  Each arm:
-///
-/// 1. Defines a method-local stack buffer `[E820Entry; 128]`
-///    (smaller than ACPI's 256 KiB — e820 entries are ~24 bytes
-///    each, so 128 entries = ~3 KiB, fits on a 16+ KiB stack).
-///    Using a stack buffer rather than a struct field keeps the
-///    `_BoardDevices` footprint small; the detected data lands in
-///    the global `E820State` via `fstart_capabilities::memory_detect`
-///    (which also persists it post-return).
-/// 2. Calls `fstart_capabilities::memory_detect(self.<field>.as_ref()?,
-///    &mut _e820_entries, <name>)` — discards the returned
-///    `(count, total)` since the global state captures both.
-/// 3. Returns `Ok(())` on success; log + halt on error.
-///
-/// Wildcard arm: log + halt.  Boards with no `MemoryDetector`
-/// device reduce to just the wildcard (dead code).
-fn memory_detect_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    let arms = enabled_indices(ctx.devices, ctx.instances, ctx.excluded)
-        .filter(|idx| device_provides(ctx, *idx, Service::MemoryDetector))
-        .map(|idx| {
-            let dev = &ctx.devices[idx];
-            let field = format_ident!("{}", dev.name.as_str());
-            let id_lit = proc_macro2::Literal::u8_unsuffixed(idx as u8);
-            let dev_name = dev.name.as_str();
-            quote! {
-                #id_lit => {
-                    // Buffer for `memory_detect()` output.  The
-                    // detected entries are also stored in the global
-                    // `E820State` for consumers (PCI host bridge,
-                    // boot protocol, CrabEFI).
-                    let mut _e820_entries =
-                        [fstart_services::memory_detect::E820Entry::zeroed(); 128];
-                    fstart_capabilities::memory_detect(
-                        self.#field
-                            .as_ref()
-                            .unwrap_or_else(|| fstart_platform::halt()),
-                        &mut _e820_entries,
-                        #dev_name,
-                    )
-                    .map_err(|_| fstart_services::device::DeviceError::InitFailed)?;
-                    Ok(())
-                }
-            }
-        });
-    quote! {
-        match id {
-            #(#arms)*
-            _ => {
-                // Executor contract violation — `memory_detect(id)` is
-                // only dispatched for ids the plan marked as
-                // `MemoryDetector` providers.
-                fstart_log::error!("memory_detect: unknown device id {}", id);
-                fstart_platform::halt();
-            }
-        }
-    }
-}
-
-/// Emit the body of `Board::acpi_prepare`.
-///
-/// Orchestrates per-device ACPI generation, matching the old
-/// `capabilities::acpi::generate_acpi_prepare` shape:
-///
-/// 1. Emits `let platform_acpi = ...;` via the shared
-///    [`capabilities::acpi::generate_platform_acpi`] helper.  This
-///    covers both the ARM SBSA and x86 flavours of
-///    [`fstart_acpi::platform::PlatformConfig`].
-///
-/// 2. For every enabled device whose driver `has_acpi` and whose
-///    instance has an `acpi_name()` set, emits a per-device
-///    `let <name>_cfg = <config struct literal>;` binding using
-///    [`config_ser::config_tokens`].  The old codegen had these
-///    bindings at `fstart_main` scope; moving them into the
-///    `acpi_prepare` body keeps them out of the rest of the adapter
-///    where they would shadow the `self.<name>` fields.
-///
-/// 3. Inside the `fstart_capabilities::acpi::prepare(&platform_acpi,
-///    |dsdt_aml, extra_tables| { ... })` closure, emits the per-device
-///    `dsdt_aml.extend(AcpiDevice::dsdt_aml(self.<name>.as_ref()?,
-///    &<name>_cfg))` + `extra_tables.extend(...)` calls.  The old
-///    generator used `&#dev_name` against a local; in the new model
-///    `self.<name>.as_ref().unwrap_or_else(|| halt())` yields the
-///    same `&T` reference.
-///
-/// 4. Includes ACPI-only device contributions (AHCI, xHCI, PCIe
-///    root) via the shared [`capabilities::acpi::generate_acpi_only_device`]
-///    emitter — these have no corresponding `self.<name>` field,
-///    so they're identical in the old and new model.
-///
-/// Boards without `config.acpi` (i.e. no `AcpiPrepare` capability)
-/// emit a `todo!()` — dead code since the executor only dispatches
-/// `AcpiPrepare` for boards where the capability is present.
-fn acpi_prepare_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    use super::capabilities::acpi as cap_acpi;
-    use super::config_ser;
-
-    // Gate on the stage actually declaring AcpiPrepare — the board
-    // may have an `acpi` RON section, but a bootblock stage that
-    // doesn’t list AcpiPrepare must not emit code referencing
-    // `fstart_acpi` / `fstart_capabilities::acpi` (those crates may
-    // not be linked into this stage’s binary).
-    let has_acpi_prepare = ctx
-        .stage
-        .capabilities
-        .iter()
-        .any(|c| matches!(c, Capability::AcpiPrepare));
-    if !has_acpi_prepare {
-        return quote! {
-            todo!("board_gen::acpi_prepare: stage does not declare AcpiPrepare")
-        };
-    }
-
-    let Some(acpi_cfg) = ctx.config.acpi.as_ref() else {
-        return quote! {
-            // No `acpi` config in RON — validation already rejected
-            // `AcpiPrepare` in the capability list, so this body is
-            // never reached at runtime.  Keep it compilable.
-            todo!("board_gen::acpi_prepare: board has no `acpi` RON config")
-        };
-    };
-
-    // Per-device config bindings, emitted as `let <name>_cfg = <literal>;`.
-    let mut config_lets = TokenStream::new();
-    // DSDT / extra_tables calls inside the closure.
-    let mut device_blocks = TokenStream::new();
-    // ACPI-only device blocks (AHCI / xHCI / PcieRoot).
-    let mut acpi_only_blocks = TokenStream::new();
-
-    for (idx, dev) in ctx.devices.iter().enumerate() {
-        let inst = &ctx.instances[idx];
-        let meta = inst.meta();
-
-        // Per-driver contributors — need both `has_acpi` and a
-        // non-None `acpi_name()` to emit anything.  Skip ACPI-only
-        // instances: they have no `self.<field>` and are handled
-        // separately below via `generate_acpi_only_device`.
-        if meta.has_acpi && inst.acpi_name().is_some() && !inst.is_acpi_only() {
-            let field = format_ident!("{}", dev.name.as_str());
-            let cfg_name = format_ident!("{}_cfg", dev.name.as_str());
-            let cfg_literal = config_ser::config_tokens(inst);
-            let drv_ty = config_ser::driver_type_tokens(inst);
-            let cfg_ty = if dev.parent.is_some() {
-                quote! { <#drv_ty as fstart_services::device::BusDevice>::Config }
-            } else {
-                quote! { <#drv_ty as fstart_services::device::Device>::Config }
-            };
-            config_lets.extend(quote! {
-                let #cfg_name: #cfg_ty = #cfg_literal;
-            });
-            device_blocks.extend(quote! {
-                dsdt_aml.extend(fstart_acpi::device::AcpiDevice::dsdt_aml(
-                    self.#field.as_ref().unwrap_or_else(|| fstart_platform::halt()),
-                    &#cfg_name,
-                ));
-                extra_tables.extend(fstart_acpi::device::AcpiDevice::extra_tables(
-                    self.#field.as_ref().unwrap_or_else(|| fstart_platform::halt()),
-                    &#cfg_name,
-                ));
-            });
-        }
-    }
-
-    // ACPI-only contributors (no `self.<name>` field — they carry
-    // their own per-variant struct literal emitted by the shared
-    // helper).  The old generator enumerates them with an independent
-    // counter so `_acpi_dev_0`, `_acpi_dev_1`, ... don't collide.
-    let mut extra_idx = 0usize;
-    for (idx, _dev) in ctx.devices.iter().enumerate() {
-        let inst = &ctx.instances[idx];
-        if !inst.is_acpi_only() {
-            continue;
-        }
-        acpi_only_blocks.extend(cap_acpi::generate_acpi_only_device(inst, extra_idx));
-        extra_idx += 1;
-    }
-
-    let platform_block = cap_acpi::generate_platform_acpi(&acpi_cfg.platform);
-    let print_hex = acpi_cfg.print_hex;
-
-    quote! {
-        #platform_block
-        #config_lets
-        self._acpi_rsdp_addr = fstart_capabilities::acpi::prepare_with_options(&platform_acpi, #print_hex, |dsdt_aml, extra_tables| {
-            #device_blocks
-            #acpi_only_blocks
-        });
-    }
-}
-
-/// Emit the body of `Board::smbios_prepare`.
-///
-/// Delegates entirely to the shared
-/// [`capabilities::smbios::generate_smbios_prepare`] helper, which is
-/// a pure function of [`BoardConfig`] — no `self.<field>` references,
-/// no per-device config bindings, no closures.  The emitted
-/// `fstart_capabilities::smbios::prepare(&SmbiosDesc { ... })` call
-/// is identical in the old and new models.
-///
-/// Boards without `config.smbios` (i.e. no `SmBiosPrepare`
-/// capability) emit a `todo!()` — dead code since the executor
-/// only dispatches `SmBiosPrepare` for boards where the capability
-/// is present.
-fn smbios_prepare_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    let has_smbios_prepare = ctx
-        .stage
-        .capabilities
-        .iter()
-        .any(|c| matches!(c, Capability::SmBiosPrepare));
-    if !has_smbios_prepare {
-        return quote! {
-            todo!("board_gen::smbios_prepare: stage does not declare SmBiosPrepare")
-        };
-    }
-    if ctx.config.smbios.is_none() {
-        return quote! {
-            // No `smbios` config in RON — validation already rejected
-            // `SmBiosPrepare` in the capability list, so this body is
-            // never reached at runtime.
-            todo!("board_gen::smbios_prepare: board has no `smbios` RON config")
-        };
-    }
-    super::capabilities::generate_smbios_prepare(ctx.config)
 }
 
 /// Emit the body of `Board::boot_media_select`.
